Fix: 컴포넌트 예시 빼기
#17
